Перенести toString(Task task) в отдельный класс CSVTaskFormat и заменить старый toString на новый в функции save. Добавить буферизацию.
