+++ src/ru/yandex/javacource/bochkareva/schedule/task/CSVTaskFormat.java
@@ -1,0 +1,13 @@
+package ru.yandex.javacource.bochkareva.schedule.task;
+
+public class CSVTaskFormat {
+
+    public static String toString(Task task) {
+        return task.getId() + ","
+                + task.getType() + ","
+                + task.getName() + ","
+                + task.getStatus() + ","
+                + task.getDescription() + ","
+                + (task.getType().equals(TaskType.SUBTASK) ? ((Subtask) task).getEpicId() : "");
+    }
+}
